Revert "Fetch routes"
